+++ src/displayer/MultipleSelect.php
@@ -72,6 +72,7 @@ class MultipleSelect extends Select
         $vars = array_merge($vars, [
             'checked' => $this->checked,
             'dataSelected' => implode(',', $dataSelected), //已经手动在后端给了选项的，不再ajax加载默认值
+            'dataDisabled' => implode(',', $this->disabledOptions),
             'select2' => $this->select2,
             'group' => $this->group,
             'options' => $this->options,